// pro-front/src/app/models/exam.ts
export interface Exam {
  _id?: string; // اختياري عند الإنشاء، مطلوب عند التعديل
  title: string; // عنوان الامتحان (مطلوب)
  description: string; // وصف الامتحان (مطلوب)
  duration: number; // مدة الامتحان بالدقائق (مطلوب)
  active: boolean; // حالة الامتحان (نشط أم لا) (مطلوب)
  created_by?: string; // معرف المستخدم الذي أنشأ الامتحان (مطلوب)
  exam_id?: string; // معرف الامتحان (اختياري، قد يكون مطلوبًا من قبل الخادم)
  createdAt?: string; // تاريخ الإنشاء (اختياري، يتم إنشاؤه تلقائيًا من قبل الخادم)
  updatedAt?: string; // تاريخ التعديل (اختياري، يتم تحديثه تلقائيًا من قبل الخادم)
}